+++ app/src/main/java/de/passwordvault/view/settings/activity_recovery/RecoveryScreen.kt
@@ -1,22 +1,22 @@
 package de.passwordvault.view.settings.activity_recovery
 
-import androidx.appcompat.widget.PopupMenu.OnDismissListener
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ExperimentalLayoutApi
+import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.text.input.rememberTextFieldState
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -29,7 +29,7 @@ import androidx.compose.material3.MenuAnchorType
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberModalBottomSheetState
@@ -38,12 +38,11 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.modifier.modifierLocalOf
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringArrayResource
@@ -52,11 +51,11 @@ import de.passwordvault.R
 import de.passwordvault.model.security.login.SecurityQuestion
 import de.passwordvault.ui.composables.BottomSheetDialog
 import de.passwordvault.ui.composables.Card
+import de.passwordvault.ui.composables.DeleteDialog
 import de.passwordvault.ui.composables.EmptyPlaceholder
 import de.passwordvault.ui.composables.Headline
 import de.passwordvault.ui.composables.TextInput
-import de.passwordvault.view.settings.dialog_security_question.SecurityQuestionDialog
-import kotlin.math.exp
+import kotlinx.coroutines.launch
 
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -105,7 +104,7 @@ fun RecoveryScreen(
                     title = stringResource(R.string.recovery_questions),
                     endIcon = painterResource(R.drawable.ic_add),
                     onClick = {
-                        //TODO...
+                        viewModel.editedSecurityQuestion = SecurityQuestion(-1, "")
                     }
                 )
             }
@@ -128,32 +127,37 @@ fun RecoveryScreen(
                 items(viewModel.securityQuestions) { securityQuestion ->
                     SecurityQuestionListRow(
                         securityQuestion = securityQuestion,
-                        onEdit = { securityQuestion ->
-                            viewModel.editedSecurityQuestion = securityQuestion
+                        onEdit = {
+                            viewModel.editedSecurityQuestion = it
                         },
-                        onDelete = { securityQuestion ->
-                            //TODO
+                        onDelete = {
+                            viewModel.deleteSecurityQuestion = it
                         }
                     )
                 }
             }
         }
         if (viewModel.editedSecurityQuestion != null) {
-            de.passwordvault.view.settings.activity_recovery.SecurityQuestionDialog(
+            SecurityQuestionDialog(
                 question = viewModel.editedSecurityQuestion!!.question,
                 answer = viewModel.editedSecurityQuestion!!.answer,
-                onQuestionChange = {
-                    viewModel.editedSecurityQuestion!!.question = it
-                },
-                onAnswerChange = {
-                    viewModel.editedSecurityQuestion!!.answer = it
-                },
+                securityQuestions = viewModel.getAvailableSecurityQuestions(),
                 onDismiss = {
                     viewModel.editedSecurityQuestion = null
                 },
-                onSave = {
-                    //TODO
-                    viewModel.editedSecurityQuestion = null
+                onSave = { question, answer ->
+                    viewModel.saveSecurityQuestion(question, answer)
+                }
+            )
+        }
+        if (viewModel.deleteSecurityQuestion != null) {
+            DeleteDialog(
+                message = stringResource(R.string.recovery_delete_message),
+                onCancel = {
+                    viewModel.deleteSecurityQuestion = null
+                },
+                onConfirm = {
+                    viewModel.deleteSecurityQuestion(viewModel.deleteSecurityQuestion!!)
                 }
             )
         }
@@ -247,6 +251,7 @@ private fun SecurityQuestionListRow(
                         )
                     },
                     onClick = {
+                        isDropdownVisible = false
                         onEdit(securityQuestion)
                     }
                 )
@@ -261,6 +266,7 @@ private fun SecurityQuestionListRow(
                         )
                     },
                     onClick = {
+                        isDropdownVisible = false
                         onDelete(securityQuestion)
                     }
                 )
@@ -270,23 +276,24 @@ private fun SecurityQuestionListRow(
 }
 
 
-@OptIn(ExperimentalMaterial3Api::class)
+@OptIn(ExperimentalMaterial3Api::class, ExperimentalLayoutApi::class)
 @Composable
 private fun SecurityQuestionDialog(
     question: Int,
     answer: String,
-    onQuestionChange: (Int) -> Unit,
-    onAnswerChange: (String) -> Unit,
+    securityQuestions: List<String>,
     onDismiss: () -> Unit,
-    onSave: () -> Unit
+    onSave: (Int, String) -> Unit
 ) {
     val sheetState = rememberModalBottomSheetState()
-    var question: Int by remember { mutableStateOf(question) }
+    val scope = rememberCoroutineScope()
+    var question: Int by remember { mutableIntStateOf(question) }
     var answer: String by remember { mutableStateOf(answer) }
+    val allQuestions: Array<String> = stringArrayResource(R.array.security_questions)
 
     BottomSheetDialog(
         sheetState = sheetState,
-        title = "SSS",
+        title = stringResource(R.string.recovery_configure_question),
         onDismiss = onDismiss,
         icon = painterResource(R.drawable.ic_edit)
     ) {
@@ -300,7 +307,17 @@ private fun SecurityQuestionDialog(
         ) {
             DropdownInput(
                 label = stringResource(R.string.recovery_question),
-                options = stringArrayResource(R.array.security_questions).toList()
+                options = securityQuestions,
+                selectedOption = if (question != -1) { stringArrayResource(R.array.security_questions)[question] } else { null },
+                onSelectionChange = {
+                    var i = 0
+                    allQuestions.forEach { q ->
+                        if (q == it) {
+                            question = i
+                        }
+                        i++
+                    }
+                }
             )
             TextInput(
                 value = answer,
@@ -309,6 +326,36 @@ private fun SecurityQuestionDialog(
                 },
                 label = stringResource(R.string.recovery_answer)
             )
+            FlowRow(
+                horizontalArrangement = Arrangement.End,
+                modifier = Modifier
+                    .align(Alignment.End)
+                    .padding(bottom = dimensionResource(R.dimen.space_vertical))
+            ) {
+                TextButton(
+                    onClick = {
+                        scope.launch {
+                            sheetState.hide()
+                        }.invokeOnCompletion {
+                            onDismiss()
+                        }
+                    }
+                ) {
+                    Text(text = stringResource(R.string.button_cancel))
+                }
+                TextButton(
+                    modifier = Modifier.padding(start = dimensionResource(R.dimen.space_horizontal_between)),
+                    onClick = {
+                        scope.launch {
+                            sheetState.hide()
+                        }.invokeOnCompletion {
+                            onSave(question, answer)
+                        }
+                    }
+                ) {
+                    Text(text = stringResource(R.string.button_save))
+                }
+            }
         }
     }
 }
@@ -318,21 +365,23 @@ private fun SecurityQuestionDialog(
 @Composable
 private fun DropdownInput(
     label: String,
-    options: List<String>
+    selectedOption: String?,
+    options: List<String>,
+    onSelectionChange: (String) -> Unit,
+    modifier: Modifier = Modifier
 ) {
     var expanded: Boolean by remember { mutableStateOf(false) }
-    val textFieldState = rememberTextFieldState(options[0])
-    var selectedOption: Int by remember { mutableIntStateOf(0) }
 
     ExposedDropdownMenuBox(
         expanded = expanded,
         onExpandedChange = {
             expanded = it
-        }
+        },
+        modifier = modifier
     ) {
         OutlinedTextField(
             readOnly = true,
-            value = options[selectedOption],
+            value = selectedOption ?: "",
             onValueChange = { },
             label = {
                 Text(label)
@@ -343,7 +392,9 @@ private fun DropdownInput(
                 )
             },
             colors = ExposedDropdownMenuDefaults.textFieldColors(),
-            modifier = Modifier.menuAnchor(MenuAnchorType.PrimaryEditable)
+            modifier = Modifier
+                .fillMaxWidth()
+                .menuAnchor(MenuAnchorType.PrimaryEditable)
         )
         ExposedDropdownMenu(
             expanded = expanded,
@@ -356,18 +407,15 @@ private fun DropdownInput(
                     text = {
                         Text(
                             text = option,
-                            style = MaterialTheme.typography.bodyMedium
+                            style = MaterialTheme.typography.bodyMedium,
+                            modifier = Modifier.padding(
+                                vertical = dimensionResource(R.dimen.space_vertical_between)
+                            )
                         )
                     },
                     onClick = {
-                        var i = 0
-                        options.forEach { o ->
-                            if (o == option) {
-                                selectedOption = i
-                            }
-                            i++
-                        }
                         expanded = false
+                        onSelectionChange(option)
                     },
                     contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding
                 )